pyproject.toml
[tool.poetry]
name = "note-guard"
version = "0.1.0"
description = ""
authors = ["Ivan Gruzdev <gruzdev.ivan.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.1"}
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
pyaspeller = "^2.0.0"
httpx = "^0.27.0"
ruff = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
